pyproject.toml
[tool.poetry]
name = "flowpredictiondemo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
numpy = "^1.26.0"
matplotlib = "^3.8.0"
torch = ">=2.0.0, !=2.0.1"
elasticai-creator = "^0.59.0"
scikit-learn = "^1.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
